- Adjust bigincrements on migrations.

+++ database/migrations/2015_01_01_150000_create_role_user_table.php
@@ -18,9 +18,9 @@ class CreateRoleUserTable extends Migration
 			'role_user',
 			function(Blueprint $table)
 			{
-				$table->increments('id');
-				$table->integer('role_id')->unsigned();
-				$table->integer('user_id')->unsigned();
+				$table->bigIncrements('id');
+				$table->bigInteger('role_id')->unsigned();
+				$table->bigInteger('user_id')->unsigned();
 				$table->timestamps();
 
 				$table->foreign('role_id')->references('id')->on('roles');
